Deployment to K8s and change RabbitMQ admin user in Go
These two steps are implemented in Go using the client-go Kubernetes
client and the Go net/http package. The third and last step is to store
the credentials of thew new admin user in AWS Secrets Manager. This can
be implemented with the AWS SDK for Go (https://github.com/aws/aws-sdk-go),
which I already tested.

However, maybe it's not worth to implement all these steps in Go, and it
makes more sense to just use a Bash script and invoke the relevant
end-user tools:

- Step 1: kubectl
- Step 2: curl
- Step 3: aws (AWS CLI)

This requires that these tools are installed on the local system, but
it's probably reasonable to assume this.

Because, even if everything is implemented in Go, there are still some
requirements: for step 1, there must be an AWS credential chain
configured on the system, aws-iam-authenticator must be installed, and
there must be a kubeconfig file. For step 3, there must be an AWS
credential chain on the system.

However, a place where it is necessary to to do these things in Go is
from the Lambda functions (post signup, etc).